Move Mutex test's threading helpers to a separate header so other tests can use them.
